[r=allenap][bug=][author=ltrager] Add comment when creating or updating tests, add missing license to stress_ng_cpu_long test, add comment to NTP test, and change NTP from network, internet to network, ntp as the NTP does not always connect to the internet.

+++ src/maasserver/api/scripts.py
@@ -192,6 +192,10 @@ class NodeScriptHandler(OperationsHandler):
             script = get_object_or_404(Script, id=int(name))
         else:
             script = get_object_or_404(Script, name=name)
+
+        if script.default:
+            raise MAASAPIValidationError("Unable to delete default script")
+
         script.delete()
         return rc.DELETED
 
@@ -292,6 +296,9 @@ class NodeScriptHandler(OperationsHandler):
         else:
             script = get_object_or_404(Script, name=name)
         try:
+            if script.default:
+                raise MAASAPIValidationError("Unable to revert default script")
+
             def gc_hook(value):
                 script.script = value
                 script.save()
